Hookup specified by Class selection. Documentation Update.
To specify the wiring type on the output end of the MegaMoto, simply declare your MegaMoto object with the right class. H-bridge = MegaMotoHB and half-bridge = MegaMotoHalfB.

There has been a lot of commenting added to all of the code I've written so far. Not enough testing has been done just yet, with the split into a base and inherited class, so this shouldn't be merged into master yet.

Just making a commit for backup purposes before I go to bed. :-)