Pre-commit script. Base class re-named.
I wanted to "package" everything much faster. Since most of my development will take place on my Windows machine, I made a CMD script. I might whip up an equivalient bash script of the same name for my Linux machine as well, but for now, this will work just fine.

The capability of re-generating documentation and then repackaging the library when bugs or code changes have been made should prove useful, I hope.

The base class name was appended with the word "Base." This should help with the documentation clarity as well as deter the temptation for users to include it in their code.